+++ assets/components/sections/index.ts
@@ -1,3 +1,5 @@
 import ResumeHeader from './ResumeHeader';
+import ResumeSections from './ResumeSections';
+import EducationSection from './EducationSection';
 
-export { ResumeHeader };
+export { ResumeHeader, ResumeSections, EducationSection };