+++ assets/components/ui/index.ts
@@ -2,5 +2,16 @@ import SectionTitle from './SectionTitle';
 import FormInput from './FormInput';
 import Button from './Button';
 import ViewHeaderData from './ViewHeaderData';
+import ViewEducationData from './ViewEducationData';
+import SaveButton from './SaveButton';
+import EditButton from './EditButton';
 
-export { SectionTitle, FormInput, Button, ViewHeaderData };
+export {
+    SectionTitle,
+    FormInput,
+    Button,
+    ViewHeaderData,
+    SaveButton,
+    ViewEducationData,
+    EditButton,
+};