+++ assets/schema/index.ts
@@ -2,6 +2,9 @@ import { yupResolver } from '@hookform/resolvers/yup';
 
 import headerSchema from './header.schema';
 
-const resumeHeaderResolver = yupResolver(headerSchema);
+import educationSchema from './education.schema';
 
-export { resumeHeaderResolver };
+const resumeHeaderResolver = yupResolver(headerSchema);
+const educationResolver = yupResolver(educationSchema);
+
+export { resumeHeaderResolver, educationResolver };